Stable removal

+++ src/Systems/TileMapSystem.cpp
@@ -38,14 +38,15 @@ TileMapSystem::TileMapSystem(std::string path, int rows, int cols)
 
 void TileMapSystem::Update(SDL_Renderer* renderer, std::unique_ptr<AssetStore>& assetstore)
 {
+    // Logger::Warning("Tiling the tilemap");
     for (auto entity : GetSystemEntities())
     {
         TileMapComponent tileMap = entity.GetComponent<TileMapComponent>();
         SDL_Texture* texture = assetstore->GetTexture(tileMap.spriteName);
 
-        for (int i = 0; i < vecs.size(); i++)
+        for (size_t i = 0; i < vecs.size(); i++)
         {
-            for (int j = 0; j < vecs[i].size(); j++)
+            for (size_t j = 0; j < vecs[i].size(); j++)
             {
                 MakeCell(tileMap, renderer, texture, vecs[i][j].second, vecs[i][j].first, j, i);
             }
@@ -53,11 +54,11 @@ void TileMapSystem::Update(SDL_Renderer* renderer, std::unique_ptr<AssetStore>& 
     }
 }
 
-void TileMapSystem::MakeCell(const TileMapComponent& tileMap, SDL_Renderer* renderer,
-                             SDL_Texture* texture, int x, int y, int locX, int locY)
+void TileMapSystem::MakeCell(const TileMapComponent& tileMap, SDL_Renderer* renderer, SDL_Texture* texture, int x,
+                             int y, int locX, int locY)
 {
     SDL_Rect src{x * tileSize, y * tileSize, tileSize, tileSize};
-    SDL_Rect dst{locX * tileScale * tileSize, locY * tileScale * tileSize,
+    SDL_Rect dst{static_cast<int>(locX * tileScale * tileSize), static_cast<int>(locY * tileScale * tileSize),
                  static_cast<int>(tileSize * tileScale), static_cast<int>(tileScale * tileSize)};
     SDL_RenderCopy(renderer, texture, &src, &dst);
 }
